% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_contrast.R
\name{assertContrast}
\alias{assertContrast}
\title{Check dimensionality of contrast matrix}
\usage{
assertContrast(design, contrast)
}
\arguments{
\item{design}{Design matrix}

\item{contrast}{Contrast matrix}
}
\value{
Side effect is used: the function stops if the ncol(design) does not equal nrow(contrast)
}
\description{
Check dimensionality of contrast matrix
}
\examples{
design <- matrix(1:20, ncol=5)
contrast <- matrix(c(-1,1,0,0,0, 0,1,0,-1,0), nrow=5)
assertContrast(design, contrast)
}
