% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_contrast.R
\name{assertDesign}
\alias{assertDesign}
\title{Check dimensionality of design matrix}
\usage{
assertDesign(nsample, design)
}
\arguments{
\item{nsample}{Integer, number of samples}

\item{design}{Design matrix}
}
\value{
Side effect is used: the function stops if sample size does not
equal ncol(matrix)
}
\description{
Check dimensionality of design matrix
}
\examples{

nsample <- 4
design <- matrix(1:20, ncol=5)
assertDesign(nsample, design)

}
