% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedaInfo.R
\name{bedaInfo}
\alias{bedaInfo}
\title{Print BEDA project information}
\usage{
bedaInfo()
}
\description{
The function is used at the end of the Rmarkdown report to print relevant information to help other colleagues finding relevant resources
}
\examples{
if(interactive()) {bedaInfo()}
}
