% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedaInfo.R
\name{biomicsPstorePath2URL}
\alias{biomicsPstorePath2URL}
\title{Translate BiOmics-Pathology pstore path to URL}
\usage{
biomicsPstorePath2URL(path)
}
\arguments{
\item{path}{Unix path}
}
\value{
Character string of biomics pstore path 
The URL is only visible inside Roche
}
\description{
Translate BiOmics-Pathology pstore path to URL
}
\examples{
biomicsPstorePath2URL("/pstore/data/biomics/")
}
