% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_dir.R
\name{createDir}
\alias{createDir}
\title{Create a directory if it does not exist, and then make sure the creation was
successful.}
\usage{
createDir(dir, showWarnings = FALSE, recursive = FALSE,
  mode = "0777")
}
\arguments{
\item{dir}{Directory name}

\item{showWarnings}{Passed to \link{dir.create}}

\item{recursive}{Passed to \link{dir.create}}

\item{mode}{Passed to \link{dir.create}}
}
\value{
Directory name (invisible)
}
\description{
The function is particularly useful for scripting.
}
\examples{

tempdir <- tempdir()
createDir(tempdir)

}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
