% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsumprop.R
\name{cumsumprop}
\alias{cumsumprop}
\title{Proportion of cumulative sum over sum}
\usage{
cumsumprop(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
the proportion cumulative sum over sum
}
\description{
Proportion of cumulative sum over sum
}
\examples{
x <- 1:4
cumsumprop(x) ## 0.1, 0.3, 0.6, 1
}
