% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{dfFactor2Str}
\alias{dfFactor2Str}
\title{Convert factor columns in a data.frame into character strings}
\usage{
dfFactor2Str(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
A data.frame with factor columns coereced into character strings
}
\description{
Convert factor columns in a data.frame into character strings
}
\examples{
exampleDf <- data.frame(Teams=c("HSV", "FCB", "FCB", "HSV"), 
             Player=c("Mueller", "Mueller", "Robben", "Holtby"), 
             scores=c(3.5, 1.5, 1.5, 1.0), stringsAsFactors=TRUE)
strDf <- dfFactor2Str(exampleDf)
stopifnot(identical(strDf[,1], c("HSV", "FCB", "FCB", "HSV")))
stopifnot(identical(exampleDf[,1], factor(c("HSV", "FCB", "FCB", "HSV"))))
}
