% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microwellPlates.R
\name{equateWellLabelWidth}
\alias{equateWellLabelWidth}
\title{Format labels for wells in microwell plates to equal widths}
\usage{
equateWellLabelWidth(wells)
}
\arguments{
\item{wells}{A vector of character strings indicating well positions, they
may be of different widths, for instance A1, A10, A12}
}
\value{
A vector of the same length, with all labels adjusted to the equal
width, with left-padding zeros added whenever it makes sense. If the input
labels are already of the same length, no change is applied.
}
\description{
Format labels for wells in microwell plates to equal widths
}
