% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{firstUp}
\alias{firstUp}
\title{Make the first alphabet of strings uppercase}
\usage{
firstUp(str)
}
\arguments{
\item{str}{A vector of character strings}
}
\value{
A vector of the same length, with the first alphabet in uppercase
}
\description{
Make the first alphabet of strings uppercase
}
\examples{
firstUp('test string')
firstUp(strsplit('many many years ago', ' ')[[1]])
}
\seealso{
\code{\link[tools]{toTitleCase}}
}
