% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openFileDevice.R
\name{getDefaultFontFamily}
\alias{getDefaultFontFamily}
\title{Get default font family}
\usage{
getDefaultFontFamily()
}
\value{
Character string, the default font family
}
\description{
Get default font family
}
