% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{lastChar}
\alias{lastChar}
\title{Return last characters from strings}
\usage{
lastChar(str)
}
\arguments{
\item{str}{A vector of character strings}
}
\value{
A vector of the same length, containing last characters
}
\description{
Return last characters from strings
}
\examples{
lastChar("Go tell it on the mountain")
lastChar(c("HSV", "FCB", "BVB"))
}
