% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{ofactor}
\alias{ofactor}
\title{Ordered factor}
\usage{
ofactor(x, ...)
}
\arguments{
\item{x}{A vector of character strings}

\item{\dots}{Other parameters passed to \code{factor}}
}
\value{
Factor with levels in the same order of the input strings.
}
\description{
Build a factor using the order of input character strings
}
\examples{

testStrings <- c("A", "C", "B", "B", "C")
(testFac <- factor(testStrings))
(testOfac <- ofactor(testStrings))

stopifnot(identical(levels(testOfac), c("A", "C", "B")))

}
\seealso{
\code{factor}
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
