% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_dir.R
\name{overwriteDir}
\alias{overwriteDir}
\title{Overwrite a directory}
\usage{
overwriteDir(dir, action = c("ask", "overwrite", "append", "no"))
}
\arguments{
\item{dir}{Chacater, path to a directory.}

\item{action}{Ask the user to input the option (\code{ask}), or one of the
following options: \code{overwrite}, \code{append}, and \code{no}. See below
for other options.}
}
\value{
If \code{action} is set to \code{overwrite}, the directory will be
deleted recursively if it exists, a new directory with the same name will be
created, and the function returns \code{TRUE}. If \code{append} is set, the
function creates the directory if necessary and returns \code{TRUE}. If
\code{no} is set, the function does nothing and returns.

If \code{action} is set to \code{ask}, user will be prompted for actions.

If \code{overwrite} is set, the directory will be removed and written anew.

If \code{append} is set, in contrast to \code{overwrite}, the directory and
 the files in it are not removed if they exists. In this case, files with the
  same name will be overwritten. Otherwise, new directories or files
  will be simply created. On the other hand, if the directory does not exist,
  it will be created.

If \code{no} is set, no action will be taken. The funciton returns
\code{FALSE}.
}
\description{
Overwrite a directory
}
\examples{
## Helper to create a test directory with files
createTestDir <- function() {
  testdir <- file.path(tempdir(), "overwriteDir_test")
  dir.create(testdir, showWarnings = FALSE)
  writeLines("First file", file.path(testdir, "file1.txt"))
  writeLines("Second file", file.path(testdir, "file2.txt"))
  return(testdir)
}
addFileToDir <- function(testdir) {
  writeLines("New file", tempfile(tmpdir=testdir))
}

## overwrite: delete the directory and create it anew
testdir <- createTestDir()
length(dir(testdir)) ## two files should be there
overwriteDir(testdir, action="overwrite")
addFileToDir(testdir)
length(dir(testdir)) ## now there should be only one file

## append: keep existing files, add new ones
overwriteDir(testdir, action="append")
addFileToDir(testdir)
length(dir(testdir)) ## now two files

## no: no action, returns FALSE
noRes <- overwriteDir(testdir, action="no")
noRes

## cleanup
unlink(testdir, recursive = TRUE)

\dontrun{
  ## ask: prompts user for action (interactive only)
  testdir <- createTestDir()
  overwriteDir(testdir, action="ask")
}

}
