% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedaInfo.R
\name{print.BEDAinfo}
\alias{print.BEDAinfo}
\title{Print BEDAinfo object}
\usage{
\method{print}{BEDAinfo}(x, ...)
}
\arguments{
\item{x}{A BEDA info object, returned by \code{\link{bedaInfo}}}

\item{...}{Ignored}
}
\value{
Invisible \code{NULL}, only side effect is used
}
\description{
Print BEDAinfo object
}
\examples{
if(interactive()) {print(bedaInfo())}
}
