% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procent.R
\name{procent}
\alias{procent}
\title{Print a decimal number in procent format}
\usage{
procent(x, fmt = "1.1")
}
\arguments{
\item{x}{a decimal number between -1 and 1}

\item{fmt}{format string, '1.1' means a digit before and after the decimal point will be printed}
}
\description{
Print a decimal number in procent format
}
\examples{
procent(c(0,0.1,0.25,1))
procent(c(0,0.1,0.25,1), fmt="1.4")
procent(c(0,-0.1,0.25,-1), fmt="+1.1")

}

