% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwdecode.R
\name{pwencode}
\alias{pwencode}
\title{Encode a password}
\usage{
pwencode(label = "VAR", key)
}
\arguments{
\item{label}{label used to encode the password}

\item{key}{password key}
}
\description{
Encode a password
}
