% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{qsystem}
\alias{qsystem}
\title{Quietly runs a system command}
\usage{
qsystem(command)
}
\arguments{
\item{command}{A system command}
}
\value{
(Invisibly) the internalized output of the command
}
\description{
Quietly runs a system command: the output is internalized and returned as an
invisible variable, and the standard error output is ignored.
}
\details{
The function runs the system command in a quiet mode. The function can be
useful in CGI scripts, for instance
}
\examples{

dateIntern <- system("date")

}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
