% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeRowsWithAnyNA.R
\name{removeColumnsWithNA}
\alias{removeColumnsWithNA}
\title{Remove columns in a matrix that contains one or more NAs}
\usage{
removeColumnsWithNA(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
A matrix, with columns containing one or more NAs removed
}
\description{
Remove columns in a matrix that contains one or more NAs
}
\examples{
myMat <- matrix(c(1:9, NA, 10:17), nrow=6, byrow=TRUE,
  dimnames=list(sprintf("R\%d", 1:6), sprintf("C\%d", 1:3)))
removeColumnsWithNA(myMat)
}
