% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeRowsWithAnyNA.R
\name{removeRowsWithNA}
\alias{removeRowsWithNA}
\title{Remove rows in a matrix that contains one or more NAs}
\usage{
removeRowsWithNA(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
A matrix, with rows containing one or more NAs removed
}
\description{
Remove rows in a matrix that contains one or more NAs
}
\examples{
myMat <- matrix(c(1:9, NA, 10:17), nrow=6, byrow=TRUE,
  dimnames=list(sprintf("R\%d", 1:6), sprintf("C\%d", 1:3)))
removeRowsWithNA(myMat)
}
