% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmat.R
\name{replaceColumnName}
\alias{replaceColumnName}
\title{Replace column names in data.frame}
\usage{
replaceColumnName(data.frame, old.names, new.names)
}
\arguments{
\item{data.frame}{A data.frame}

\item{old.names}{Old column names to be replaced}

\item{new.names}{New column names}
}
\value{
Data.frame with column names updated
}
\description{
Replace column names in data.frame
}
\examples{

clubs <- data.frame(Points=c(21,23,28,24), Name=c("BVB", "FCB", "HSV",
"FCK"), games=c(12,11,11,12))
replaceColumnName(clubs, c("Points", "games"), c("Punkte", "Spiele"))

}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
