% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{ribiosTempdir}
\alias{ribiosTempdir}
\title{A temporary directory which (1) every machine in the cluster has access to and (2) has sufficient space}
\usage{
ribiosTempdir()
}
\value{
a character string of the directory name
}
\description{
A temporary directory which (1) every machine in the cluster has access to and (2) has sufficient space
}
\seealso{
\code{\link{ribiosTempfile}}
}

