% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{ribiosTempfile}
\alias{ribiosTempfile}
\title{A temporary file which (1) every machine in the cluster has access to and (2) there is sufficient space}
\usage{
ribiosTempfile(pattern = "file", tmpdir = ribiosTempdir(), fileext = "")
}
\arguments{
\item{pattern}{Character string, file pattern}

\item{tmpdir}{Character string, temp directory}

\item{fileext}{CHaracter string, file name extension (suffix)}
}
\value{
a character string of the file name
}
\description{
A temporary file which (1) every machine in the cluster has access to and (2) there is sufficient space
}
\seealso{
\code{\link{ribiosTempdir}}
}
