% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosUtils.R
\docType{package}
\name{ribiosUtils-package}
\alias{ribiosUtils-package}
\title{ribiosUtils: Utilities from and Interface to the Bioinfo-C (BIOS) Library}
\description{
Provides interface to the Bioinfo-C (internal name: BIOS) library and utilities. 'ribiosUtils' is a Swiss-knife for computational biology in drug discovery, providing functions and utilities with minimal external dependency and maximal efficiency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bedapub/ribiosUtils}
  \item \url{https://bedapub.github.io/ribiosUtils/}
  \item Report bugs at \url{https://github.com/bedapub/ribiosUtils/issues}
}

}
\author{
\strong{Maintainer}: Jitao David Zhang \email{jitao_david.zhang@roche.com} (\href{https://orcid.org/0000-0002-3085-0909}{ORCID}) [contributor]

Authors:
\itemize{
  \item Clemens Broger [contributor]
}

Other contributors:
\itemize{
  \item F.Hoffmann-La Roche AG [copyright holder]
  \item Junio C Hamano [copyright holder]
  \item Jean Thierry-Mieg [copyright holder]
  \item Konrad Rudolph [copyright holder]
  \item Richard Durbin [copyright holder]
}

}
\keyword{internal}
