% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosUtils.R
\docType{package}
\name{ribiosUtils}
\alias{ribiosUtils}
\alias{ribiosUtils-package}
\title{ribiosUtils}
\description{
ribiosUtils is a swiss-knife package providing misc utilities
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>, with inputs from
   Clemens Broger, Martin Ebeling, Laura Badi and Roland Schmucki
}
