% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowscale.R
\name{rowscale}
\alias{rowscale}
\title{S3 method for row-scaling}
\usage{
rowscale(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{Any object}

\item{center}{Logical, whether centering should be done before scaling}

\item{scale}{Logical, whether scaling should be done}
}
\description{
S3 method for row-scaling
}
