% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrank.R
\name{rrank.matrix}
\alias{rrank.matrix}
\title{Get reverse rank orders in each column}
\usage{
\method{rrank}{matrix}(x, ...)
}
\arguments{
\item{x}{A matrix}

\item{...}{Passed to \code{\link{rank}}}
}
\value{
A matrix of the same dimension and attributes of the input matrix, 
  with reverse rank orders of each column
}
\description{
Get reverse rank orders in each column
}
\examples{
testMatrix <- matrix(c(3,6,4,5,2,4,8,3,2,5,4,7), ncol=3, byrow=FALSE)
rrank(testMatrix)
}
