% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrank.R
\name{rrank}
\alias{rrank}
\alias{rrank.default}
\title{Reverse rank order}
\usage{
rrank(x, ...)

\method{rrank}{default}(x, ...)
}
\arguments{
\item{x}{A numeric, complex, character or logical vector}

\item{...}{Passed to \code{\link{rank}}}
}
\value{
A vector of numbers of the same length as the input, 
    giving reverse rank orders.
    
The function returns the reverse rank order, i.e. in the descending order
}
\description{
Reverse rank order
}
\examples{
testVec <- c(3,6,4,5)
rank(testVec)
rrank(testVec)
}
\seealso{
\code{\link{rank}}
}
