% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{rsetdiff}
\alias{rsetdiff}
\title{Reverse setdiff}
\usage{
rsetdiff(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{another vector}
}
\value{
Similar to setdiff, but with elements in y but not in x
}
\description{
Reverse setdiff
}
\details{
reverse setdiff, i.e. rsetdiff(x,y) equals setdiff(y,x)
}
\examples{
testVec1 <- LETTERS[3:6]
testVec2 <- LETTERS[5:7]
rsetdiff(testVec1, testVec2)
}
\author{
Jitao David Zhang
}
