% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueNonNA.R
\name{uniqueNonNA}
\alias{uniqueNonNA}
\title{Make a vector free of NA and unique}
\usage{
uniqueNonNA(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A unique vector without NA
}
\description{
Make a vector free of NA and unique
}
\examples{

testVec <- c(3,4,5,NA,3,5)
uniqueNonNA(testVec)

}
