% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedaInfo.R
\name{bedaInfo}
\alias{bedaInfo}
\title{Print BEDA project information}
\usage{
bedaInfo()
}
\value{
A list, including pstore path, URL, git address, and user id
The function is used at the end of the Rmarkdown report to print relevant information to help other colleagues finding relevant resources
}
\description{
Print BEDA project information
}
\examples{
if(interactive()) {bedaInfo()}
}
