% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLog.R
\name{closeLoggerConnections}
\alias{closeLoggerConnections}
\title{Close connections to all loggers 
This function closes all open connections set up by loggers
It is automatically run at the end of the R session (setup by \code{\link{registerLog}})}
\usage{
closeLoggerConnections()
}
\description{
Close connections to all loggers 
This function closes all open connections set up by loggers
It is automatically run at the end of the R session (setup by \code{\link{registerLog}})
}
\seealso{
\code{\link{registerLog}}
}
