% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardIndex.R
\name{jaccardIndex}
\alias{jaccardIndex}
\title{Calculate the Jaccard Index between two vectors}
\usage{
jaccardIndex(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
The Jaccard Index, a number between 0 and 1
}
\description{
Calculate the Jaccard Index between two vectors
}
\examples{
myX <- 1:6
myY <- 4:9
jaccardIndex(myX, myY)

myX <- LETTERS[1:5]
myY <- LETTERS[1:10]
jaccardIndex(myX, myY)
}
