% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pScore.R
\name{pScore}
\alias{pScore}
\title{Transform p-values to a continuous score}
\usage{
pScore(p, sign = 1)
}
\arguments{
\item{p}{\emph{p}-value between (0,1]}

\item{sign}{Sign of the score, either positive (in case of positive
numbers), negative (in case of negative numbers), or zero.  In case a
logical vector, \code{TRUE} is interpreted as positive and \code{FALSE} is
interpreted as negative.}
}
\description{
The function maps p value between 0 and 1 to a continuous score by the
following equation: abs(log10(p))*sign.
}
\examples{

testPvals <- c(0.001, 0.01, 0.05, 0.1, 0.5, 1)
pScore(testPvals)
testPvalSign <- rep(c(-1,1), 3)
pScore(testPvals, sign=testPvalSign)
testLog <- rep(c(TRUE, FALSE),3)
pScore(testPvals, testLog)

}
