% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{reload}
\alias{reload}
\title{Reload a package}
\usage{
reload(pkg)
}
\arguments{
\item{pkg}{Character string, name of the package}
}
\value{
Side effect is used.
}
\description{
Reload a package by first detaching and loading the library.
}
\note{
So far only character is accepted
}
\examples{

\dontrun{
library(ribiosUtils)
reload(ribiosUtils)
}

}
\seealso{
\code{\link{detach}} and \code{\link{library}}
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
