% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsumprop.R
\name{sortAndFilterByCumsumprop}
\alias{sortAndFilterByCumsumprop}
\alias{\code{eightytwenty}}
\alias{This}
\alias{function}
\alias{can}
\alias{be}
\alias{useful}
\alias{to}
\alias{extract}
\alias{from}
\alias{a}
\alias{long}
\alias{numeric}
\alias{vector}
\alias{the}
\alias{largest}
\alias{items}
\alias{that}
\alias{dominate}
\alias{sum}
\alias{of}
\title{Sort a numeric vector and filter by a threshold of cumsumprop}
\usage{
sortAndFilterByCumsumprop(x, thr = 0.9)
}
\arguments{
\item{x}{Numeric vector, usually named}
}
\value{
Another numeric vector, likely shorter than x, with only items whose \code{cumsumprop} is equal or lower than \code{thr}. The rest items are summed into one new item, with the name \code{rest}
}
\description{
Sort a numeric vector and filter by a threshold of cumsumprop
}
\examples{
x <- c("A"=1,"B"=2,"C"=3,"D"=4,"E"=400,"F"=500)
sortAndFilterByCumsumprop(x, thr=0.99) ## F and E should be returned
}
