% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubborngc.R
\name{stubborngc}
\alias{stubborngc}
\title{Repeat garbage-collecting until all resource is freed}
\usage{
stubborngc(verbose = FALSE, reset = TRUE)
}
\arguments{
\item{verbose}{Logical, verbose or not}

\item{reset}{Logical, reset or not.}
}
\value{
Side effect is used.
}
\description{
stubborngc repeats collecting garbage untill no more resource can be freed
}
\examples{

stubborngc()

}
\seealso{
\code{\link{gc}}
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
\keyword{~kwd2}
